Move blocks/header to components
