+++ src/pages/index.js
@@ -7,7 +7,7 @@ import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { ClientProvider } from 'context/ClientContext'
 import get from 'utils/get'
-import Header from 'blocks/Header'
+import Header from 'components/Header'
 import Filters from 'blocks/Filters'
 import MoviesList from 'blocks/MoviesList/index'
 import { LANGUAGES } from 'data/constants'
